Version .9.5.8
Removes use of the DescribeStream API for calculations, which commonly resulted in throttling
Makes the use of the UpdateShardCount API the default for both manual and autoscaling
Adds a waitForCompletion option (set to true by default) which blocks operation until the stream transitions back to Active status. This results in much faster autoscaling but reduces the log trace when scaling is done.