Clear up wording around scalePct
In the existing readme, the behaviour of `scalePct` wasn't clear - would setting it to `115%` increase our shard count by 15%, or would it slightly more than double our capacity? This commit adds an explanation to the behaviour of `scalePct`, and also removes the `scalePct: 100` value from the example configuration, which if used as is would cause the autoscaler to crash, as 100% is an invalid scaling percentage.